ml-model/ml-model/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pillow
numpy
tensorflow
python-dotenv==1.0.0
